Organise files better
