+++ app/javascript/components/bootcamp/CSSExercisePage/checks/css/elementHasProperty.ts
