+++ app/javascript/components/bootcamp/CSSExercisePage/checks/css/exactPropertiesUsed.ts
