+++ app/javascript/components/bootcamp/CSSExercisePage/checks/runCssChecks.ts
@@ -1,10 +1,8 @@
-import { tagOccursNoMoreThan } from './tagOccursNoMoreThan'
-
-export type HtmlCheck = {
-  function: string
-  matcher: 'toBeTrue' | 'toBeFalse'
-  errorHtml: string
-}
+import {
+  elementHasProperty,
+  elementHasPropertyValue,
+} from './css/elementHasProperty'
+import { exactPropertiesUsed } from './css/exactPropertiesUsed'
 
 export type CheckResult = {
   result: boolean | null
@@ -18,7 +16,9 @@ export type ChecksResult = {
 }
 
 const checkFunctions: Record<string, Function> = {
-  tagOccursNoMoreThan,
+  elementHasProperty,
+  elementHasPropertyValue,
+  exactPropertiesUsed,
 }
 
 function evaluateMatch(result: boolean, matcher: string): boolean {
@@ -31,11 +31,7 @@ function evaluateMatch(result: boolean, matcher: string): boolean {
       throw new Error(`Unimplemented matcher: ${matcher}`)
   }
 }
-
-export function runHtmlChecks(
-  checks: Check[],
-  htmlValue: string
-): ChecksResult {
+export function runChecks(checks: Check[], cssValue: string): ChecksResult {
   const results: CheckResult[] = checks.map((check) => {
     try {
       const funcMatch = check.function.match(/([a-zA-Z0-9_]+)\((.*)\)/)
@@ -50,17 +46,20 @@ export function runHtmlChecks(
 
       try {
         const safe_eval = eval
-        args = safe_eval(`(${argsString})`)
+        args = safe_eval(argsString)
       } catch (error) {
+        // TODO: show this only in dev mode
         throw new Error(`Invalid arguments format: ${argsString}`)
       }
 
       const func = checkFunctions[funcName]
       if (!func) {
+        // TODO: show this only in dev mode
         throw new Error(`Function not found: ${funcName}`)
       }
 
-      const result = func(htmlValue, args)
+      const result = func(cssValue, args)
+
       const passes = evaluateMatch(result, check.matcher)
 
       return {
